v1.0.1 - 2023 Updates to the Windows side
* v1.0.1 - 2023 Updates to the Windows side

-*-*-*-*-*--*--***--
The UWP file erasure experiment is proving to be quite a disaster to get stable...so maybe need to re-write this back into old school WPF. The problem there is that Lottie doesn't function, so we lose the neat animations that inspired this project. Definitely a bit of a dilemma.
Alas the Windows side is proving more to be a "how-NOT-to-build" something rather than How-To ;) 
The project still only works in the debugger mostly, and doesn't seem able to deploy due to the utter bizarre UWP deployment process. WPF is the clear winner here, even after 15+ years. I was hoping there would be an easy Windows native language with advanced C# support (for things like Lottie), but alas, it is not yet so. Microsoft loves their old languages. 
-*-*-*-*-*--*--***--

* Fixed lots of bugs and upgraded the AppService component substantially

- Needed more async await operators in the WPF project to keep things sync'd tightly

- Add lots more status feedback from the WPF->UWP side

- Upgrade the byte pattern to be static allocated once at startup, switch to fixed 4096 size. Fixed bugs in writing remainders to files.

- Add force read-only flag removal

- Lots of animation loop updates, some UI tweaks, etc.

* Fix more critical bugs regarding await async on the Main thread

- Use Low priority onm the Dispatcher.RunAsync methods helps alot
- Don't try to update an ever expanding string object, performance is very poor in UWP

* Add progress bar to know where current file status is.

- Add different blackhole play rates based on whether idle or actively erasing
- Add read-only flag stripping to WPF